requirements.txt
langchain==1.2.0
langchain-core
langchain-community
langchain-openai

python-dotenv
fastapi
pypdf
faiss-cpu
jinja2
uvicorn